### 2019-07-15 #1569 @whedon
[PRE REVIEW]: SG-t-SNE-Π: Swift Neighbor Embedding of Sparse Stochastic Graphs
**Submitting author:** @pitsianis (<a href="http://orcid.org/0000-0002-7353-3524">Nikos Pitsianis</a>)
**Repository:** <a href="https://github.com/fcdimitr/sgtsnepi" target ="_blank">https://github.com/fcdimitr/sgtsnepi</a>
**Version:** v1.0.0
