### 2019-07-15 09e20cd
Update paper.md
