### 2019-07-15 #1569 @arfon
:wave: @pitsianis  - this PR should fix your paper compilation: https://github.com/fcdimitr/sgtsnepi/pull/1
