### 2019-07-15 #1569 @labarba
> 👋 @labarba - would you be able to edit this submission for JOSS?

I could try, although I am not familiar with graph analysis. This looks like heavy duty computer science, and I might have a hard time finding reviewers. Could @jedbrown or @poulson pitch in here, to see if they would be in a better to edit?
