### 2019-07-15 #1569 @labarba
@pitsianis — Could you take a stab at improving the paper's readability for a non-specialist? Note that our [documentation](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) requests that you include:

- a summary describing high-level functionality and purpose of the software for a diverse, non-specialist audience;
- a clear statement of need that illustrates the research application of the software (including references to other software meeting similar needs).
