### 2019-07-15 #1569 @pitsianis
Hello there,

Thank you @afron for fixing the pdf generation.

Dear @labarba, I would like to propose Rio Yokota (rioyokota@gsic.titech.ac.jp) as a reviewer. He is very familiar with the methodologies and prior art in this field (N-body, Barnes Hut method) and he will appreciate our improvements.
