### 2019-07-16 #1569 @pitsianis
> @pitsianis — Could you take a stab at improving the paper's readability for a non-specialist?

Thank you @labarba for the quick feedback. We have rewritten the summary accordingly. Your comments are greatly appreciated.
