### 2019-07-16 928a9ff
Fixed issue with conventional t-SNE support & FLANN
