### 2019-07-16 2c32152
Updated documentation
