### 2019-07-17 #1577 @labarba
👋  @tingyu66, @rioyokota — Thank you for agreeing to review this JOSS submission! This is where the action happens. You both have a review checklist to work your way through. Feel free to ask any questions here. You can post your reviewer comments on this thread and open issues in the software repository as needed (add a cross-link to those here). Godspeed!
