### 2019-07-23 #1577 @labarba
👋 @pitsianis — Our reviewers are asked to check the following item:

> Authorship: Has the submitting author (@pitsianis) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

We realize that commits are not necessarily a full representation of contributions to a software project, but they do provide a visible signal for them. Therefore, if absent, we request that authors confirm that they are aware of the JOSS [authorship](https://joss.readthedocs.io/en/latest/submitting.html#authorship) policy, and conform to it.
