### 2019-07-23 #1577 @pitsianis
Yes, I confirm that all authors in the JOSS paper submission have contributed to the design, analysis and the implementation of the associated software release in substantial ways. All co-authors agree to be listed as such, agree to be accountable for all aspects of the work, and to notify JOSS if any retraction or correction of mistakes is needed after publication.

Do you need such a declaration from each co-author separately, or in a special form?
