### 2019-07-23 #1577 @rioyokota
paper.md does not contain list of authors.
