### 2019-07-23 #1577 @rioyokota
paper.md does not have references. README.md does but with no DOI for any of the references.
