### 2019-07-23 #1577 @danielskatz
Hi @rioyokota - re your last two comments, please see the pdf generated from the md and bib files, a few comments up (https://github.com/openjournals/joss-papers/blob/joss.01577/joss.01577/10.21105.joss.01577.pdf)
