### 2019-07-23 #1577 @rioyokota
Yes, the authors, references, and DOI are available in the generated pdf.
