### 2019-07-23 ed81c04
Local changes, during review process
updated README.pdf -- added doxygen in dependencies
