### 2019-07-23 #1577 @pitsianis
Thank you all for the effort, you are inspiring me to volunteer as a reviewer as well.

We have discovered an embarrassing typo and a couple of minor ones on the file paper.md. Shall we push an update or better wait for the review process to be completed?
