### 2019-07-24 #1577 @labarba
@rioyokota — does this comment require action, in your view?

> The installation instructions require sudo access.
