### 2019-07-24 #1577 @rioyokota
What has been the stance of JOSS in the past? Many people working on shared systems (supercomputers) will not be able to install this software because they cannot simply ask the admin to install the dependencies. If there were instructions on how to install these dependencies locally, that would cover a much broader audience.
