### 2019-07-24 #1577 @labarba
JOSS has no express stance on this. I see a recently published [paper](https://doi.org/10.21105/joss.01531) that also uses a `sudo` installation, but the use context there is different. It perhaps depends on how the software will likely be used. If this software will most likely be used in shared HPC scenarios, it could make sense to suggest a different installation process. Or the authors could see to provide a Singularity container image?
