### 2019-07-24 #1577 @pitsianis
We do state that our primary target is the personal computer and laptop.

This issue is really with the packages we depend upon and their support for user space builds (and their dependencies).

The container solution is a clean way to deal with this issue, especially if we can provide a script to put together the image.
