### 2019-07-24 0df0c84
Typo corrections
