+++ paper.md
@@ -35,22 +35,22 @@ SG-t-SNE-$\Pi$ is a high-performance software for swift embedding of a
 large, sparse, stochastic graph/network into a $d$-dimensional space
 ($d = 1,2,3$) on a shared-memory computer, especially on personal laptop
 and desktop computers. Graphs/networks are an important type of
-relational data. They arise ubiquitously in real-world applications and
-various research fields, such as biological networks, social networks,
-communication networks, food webs, word co-occurrence networks; see
+relational data, arising ubiquitously in real-world applications and
+various research fields. Such data include biological networks, social networks,
+communication networks, food webs, word co-occurrence networks, etc; see
 [@Kovacs2019; @Yang2015] for more real-world networks. Graph embedding
 maps each vertex of the graph to a $d$-dimensional feature vector. Graph
-embedding to a $d$-dimensional space with $d=1,2,3$ is frequently used
+embedding into a $d$-dimensional space with $d=1,2,3$ is frequently used
 in data-based scientific studies for visual inspection of data,
 interpretation of network-based analysis results, interactive inquiries
 and hypothesis generation.
 
 The software SG-t-SNE-$\Pi$ and its underlying algorithm are built upon
 precursor algorithms and software for stochastic neighbor embedding of
-high-dimensional data, in particular, the original Stochastic Neighbor
-Embedding (SNE) algorithm by Hinton and van der Maaten [@Hinton2003],
-the algorithm for t-Distributed Stochastic Neighbor Embedding (t-SNE) by
-van der Maaten [@Maaten2008], and their
+high-dimensional data, namely the original Stochastic Neighbor
+Embedding (SNE) algorithm by Hinton and Roweis [@Hinton2003],
+the algorithm for t-distributed Stochastic Neighbor Embedding (t-SNE) by
+van der Maaten and Hinton [@Maaten2008], and their
 variants [@VanDerMaaten2014; @Linderman2019].[^1][^2] The t-SNE
 algorithm has successfully assisted scientific discoveries, as reported
 in numerous articles in Nature and Science magazines. However, previous
@@ -58,26 +58,26 @@ t-SNE algorithms and software are limited in two aspects:
 (i) The algorithms require that the data points be in a metric space and the
 associated graph (internally generated) be regular with a constant
 degree. In many real-world networks, the vertices do not readily reside
-in a metric space, their degrees of connections vary greatly, far from
+in a metric space, and their degrees vary greatly, far from
 constant.
 (ii) The software is limited in practical use either to small graphs/networks
-or to embedding of $d<3$ dimensions.
-We remove these limitations. SG-t-SNE-$\Pi$ admits arbitrary large,
-sparse, stochastic graphs/networks. It is used in [@Pitsianis2019] for
+or to embedding into $d<3$ dimensions.
+We remove both limitations. SG-t-SNE-$\Pi$ admits arbitrary large,
+sparse, stochastic graphs/networks. It is demonstrated in [@Pitsianis2019] for
 novel, autonomous embedding of large, real-world stochastic networks.
 SG-t-SNE-$\Pi$ also enables fast three-dimensional (3D) graph embedding,
 which preserves and reveals more or even critical structural information
 as shown in [@Pitsianis2019], on modern laptop and desktop computers
 with ease of use.
 
-SG-t-SNE-$\Pi$ is implemented in `C++`. It takes as input a stochastic
+SG-t-SNE-$\Pi$ is implemented in C++. It takes as input a stochastic
 graph and outputs $d$-dimensional coordinate vectors, one for each
-vertex. In addition, we provide two particular types of interfaces. The
+vertex. We provide two additional interfaces. The
[…]
