+++ references.bib
@@ -1,4 +1,3 @@
-
 @article{VanDerMaaten2014,
   title = {Accelerating T-{{SNE}} Using Tree-Based Algorithms},
   volume = {15},
@@ -21,7 +20,7 @@
 }
 
 @article{Blumofe1996,
-  title = {Cilk: {{An}} Efficient Multithreaded Runtime System},
+  title = {Cilk: An Efficient Multithreaded Runtime System},
   volume = {37},
   number = {1},
   journal = {Journal of Parallel and Distributed Computing},
@@ -44,7 +43,7 @@
 
 @inproceedings{Buluc2009,
   title = {Parallel Sparse Matrix-Vector and Matrix-Transpose-Vector Multiplication Using Compressed Sparse Blocks},
-  booktitle = {Proceedings of {{Annual Symposium}} on {{Parallelism}} in {{Algorithms}} and {{Architectures}}},
+  booktitle = {Proceedings of the 21st {{Annual Symposium}} on {{Parallelism}} in {{Algorithms}} and {{Architectures}}},
   doi = {10.1145/1583991.1584053},
   author = {Bulu{\c c}, Aydin and Fineman, Jeremy T. and Frigo, Matteo and Gilbert, John R. and Leiserson, Charles E.},
   year = {2009},
@@ -53,10 +52,10 @@
 
 @inproceedings{Pitsianis2019,
   title = {Spaceland Embedding of Sparse Stochastic Graphs},
-  booktitle = {Proceedings of {{IEEE High Performance Extreme Computing Conference}}},
+  booktitle = {{{IEEE High Performance Extreme Computing Conference}}},
   author = {Pitsianis, Nikos and Iliopoulos, Alexandros-Stavros and Floros, Dimitris and Sun, Xiaobai},
   year = {2019},
-  note = {(to appear)}
+  note = {(To appear)}
 }
 
 @article{Yang2015,
@@ -85,6 +84,7 @@
 
 @inproceedings{Hinton2003,
   title = {Stochastic Neighbor Embedding},
+  volume = {15},
   booktitle = {Advances in {{Neural Information Processing Systems}}},
   author = {Hinton, Geoffrey E. and Roweis, Sam T.},
   year = {2003},
