### 2019-07-24 fcdimitr/sgtsnepi#2 @tingyu66
failed to link on macOS
Hello @pitsianis,

I was trying to configure and compile the code on my Mac running macOS `10.14.6`. I used Intel compiler icpc `19.0.4.233`, and the dependencies are installed in `/usr/local/lib`.

Since `/usr/local/lib` is not a default library search path for `ld` on macOS, I had to add it to `LDFLAGS` and configured as below:
```
 ./configure CXX=icpc LDFLAGS=-L/usr/local/lib
```
It went through and confirmed all the dependencies. Then when I compiled with
```
make all
```
it gave me a linking error
```
ld: library not found for -lflann
```
it seems that the variable `LDFLAGS` was not passed to Makefile, I have to manually add `-L/usr/local/lib` to the variable `LIBS` in the Makefile in order to compile. Could you take a look?
