### 2019-07-24 #1577 @tingyu66
I had a linking error when trying to compile on macOS and opened an issue here: https://github.com/fcdimitr/sgtsnepi/issues/2.
