### 2019-07-24 fcdimitr/sgtsnepi#2 @pitsianis
@tingyu66 please try the branch joss-review to confirm that now the issue is fixed.
