### 2019-07-26 #1577 @labarba
@tingyu66, @rioyokota — Thank you both for completing your review checklist. To move forward with accepting this submission, could we have a statement from you here summarizing your assessment of the software, and giving your recommendation to publish?
