### 2019-07-28 #1577 @rioyokota
The software is easy to use and highly relevant for the field of machine learning. I recommend publication without further modification.
