### 2019-07-30 #1577 @tingyu66
The documentation is easy to follow and the examples (in both the code and the supplementary material) clearly demonstrate some suitable embedding applications. I recommend this paper for publication.
