### 2019-07-30 #1577 @labarba
@pitsianis — you have a handful of in-text citations, meaning that the citation is a part of the sentence, with extra parens: "see (Smith 2000)" or "shown in (Jones 2001)." In these cases, you need to use the correct [citation syntax](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax) to get rid of the parens.
