### 2019-07-30 #1577 @labarba
Could you add a link to the arXiv preprint for this reference?
https://arxiv.org/abs/1906.05582
