+++ references.bib
@@ -55,6 +55,7 @@
   booktitle = {{{IEEE High Performance Extreme Computing Conference}}},
   author = {Pitsianis, Nikos and Iliopoulos, Alexandros-Stavros and Floros, Dimitris and Sun, Xiaobai},
   year = {2019},
+  url = {https://arxiv.org/abs/1906.05582},
   note = {(To appear)}
 }
 
