### 2019-07-30 2b79207
Update paper.md
in-text citations
