### 2019-07-30 75b8257
Update references.bib
Fixed Hinton's middle initial and added (to appear) in our conference paper
