+++ references.bib
@@ -36,7 +36,7 @@
   volume = {9},
   number = {Nov},
   journal = {Journal of Machine Learning Research},
-  author = {{van der Maaten}, Laurens and Hinton, Geoffrey},
+  author = {{van der Maaten}, Laurens and Hinton, Geoffrey E.},
   year = {2008},
   pages = {2579-2605}
 }
@@ -52,7 +52,7 @@
 
 @inproceedings{Pitsianis2019,
   title = {Spaceland Embedding of Sparse Stochastic Graphs},
-  booktitle = {{{IEEE High Performance Extreme Computing Conference}}},
+  booktitle = {{{IEEE High Performance Extreme Computing Conference (To appear)}}},
   author = {Pitsianis, Nikos and Iliopoulos, Alexandros-Stavros and Floros, Dimitris and Sun, Xiaobai},
   year = {2019},
   url = {https://arxiv.org/abs/1906.05582},
