### 2019-07-30 fb31f61
Update paper.md
in-line citation
