### 2019-07-30 8fa86d3
Update paper.md
typos
