### 2019-07-30 #1577 @labarba
Looking good! When you're ready, the next steps are:

- [ ] Make a new release of your software, and list the version tag of the archived version here.
- [ ] Check the Zenodo deposit has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID to the metadata.
- [ ] Please list the Zenodo DOI of the archived version here.
