### 2019-07-30 1922bf5
Update paper.md
typo
