### 2019-07-30 #1577 @pitsianis
@labarba Thank you for your help and the timely and gentle nudges to get all our ducks in a row.

A new release is made (v1.0.1).
The Zenodo DOI with this release archived is 10.5281/zenodo.3355679
