### 2019-07-30 #1577 @labarba
Could you make the title of the Zenodo deposit match the title for this JOSS submission? This can be done editing the metadata (no need for a new DOI version).
