### 2019-07-30 #1577 @pitsianis
@labarba -- We corrected the title of the Zenodo deposit to match the title for this JOSS submission.
