### 2019-07-31 #1577 @labarba
Huge thanks to  @tingyu66, @rioyokota for reviewing this submission—JOSS is grateful for your contribution 🙏
